Updated Journal Entry policy for delete action

+++ app/Http/Controllers/JournalEntriesController.php
@@ -28,8 +28,7 @@ class JournalEntriesController extends Controller
      */
     public function index()
     {
-        $user_id = auth()->user()->id;
-        $journal_entries = JournalEntry::where('user_id', $user_id)->orderBy('created_at', 'desc')->paginate(5);
+        $journal_entries = auth()->user()->journal_entries()->orderBy('created_at', 'desc')->paginate(5);
         return view('journal.index')->with('journal_entries', $journal_entries);
     }
 
@@ -89,11 +88,6 @@ class JournalEntriesController extends Controller
             return redirect('/journal')->with('error', 'Unauthorized Page');
         }
 
-        /* // Check for correct user
-        if(auth()->user()->id !== $journal_entry->user_id){
-            return redirect('/journal')->with('error', 'Unauthorized Page');
-        } */
-
         return view('journal.edit')->with('journal_entry', $journal_entry);
     }
 
@@ -127,6 +121,11 @@ class JournalEntriesController extends Controller
     public function destroy($id)
     {
         $journal_entry = JournalEntry::findOrFail($id);
+
+        if (auth()->user()->cant('delete', $journal_entry)) {
+            return redirect('/journal')->with('error', 'Unauthorized Page');
+        }
+
         $journal_entry->delete();
         return redirect('/journal')->with('success', 'Journal Entry Deleted');
     }
